modules/policy/functions/requirements.txt
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-bigquery==3.11.4
google-cloud-compute==1.14.1
google-cloud-sql==3.4.4
google-cloud-logging==3.8.0
functions-framework==3.4.0